+++ spec/services/export/discord/insights/payload_service_spec.rb
@@ -12,6 +12,23 @@ describe Export::Discord::Insights::PayloadService, type: :service do
       expect(service_call.result.size.positive?).to be_truthy
       expect(service_call.success?).to be_truthy
     end
+
+    context 'when insightable is unaccessable' do
+      before do
+        insightable.update!(accessable: false)
+      end
+
+      it 'renders message about unaccessability', :aggregate_failures do
+        service_call
+
+        expect(service_call.result.size.positive?).to be_truthy
+        expect(
+          service_call.result
+            .include?("**#{insightable.class.name} #{insightable.title} has access error**\n")
+        ).to be_truthy
+        expect(service_call.success?).to be_truthy
+      end
+    end
   end
 
   context 'with insights' do