+++ spec/services/export/slack/insights/payload_service_spec.rb
@@ -12,6 +12,19 @@ describe Export::Slack::Insights::PayloadService, type: :service do
       expect(service_call.result[:blocks].size).to eq 2
       expect(service_call.success?).to be_truthy
     end
+
+    context 'when insightable is unaccessable' do
+      before do
+        insightable.update!(accessable: false)
+      end
+
+      it 'renders message about unaccessability', :aggregate_failures do
+        service_call
+
+        expect(service_call.result[:blocks].size).to eq 3
+        expect(service_call.success?).to be_truthy
+      end
+    end
   end
 
   context 'with insights' do